moved camera creation
- camera and spring arm are now created in PlayerCharacter as not all characters need them
